requirements.txt
openai>=1.41.0
requests>=2.32.0
beautifulsoup4>=4.12.2
lxml>=5.2.2
python-dotenv>=1.0.1
flask>=3.0.0
pydantic>=2.8.0
tenacity>=8.3.0
playwright>=1.54.0
readability-lxml>=0.8.4

